requirements.txt
# Install with: pip install --only-binary=:all: -r requirements.txt
# (binary-only avoids needing MSVC build tools on Windows for chromadb / hnswlib)

# Core RAG stack
chromadb>=1.0,<2.0
sentence-transformers>=3.0
rank-bm25==0.2.2

# Document parsing
pypdf>=5.0
python-docx>=1.1

# Data / utilities
numpy>=1.26
pandas>=2.0
tqdm>=4.66

# Notebook + plotting
jupyter>=1.0
matplotlib>=3.8
seaborn>=0.13

# LLM reasoning (OpenRouter via OpenAI-compatible client)
openai>=1.40
python-dotenv>=1.0
